+++ pygam/__init__.py
@@ -1,0 +1,7 @@
+"""
+GAM toolkit
+"""
+
+__version__ = '0.2.9'
+
+from pygam import *
