+++ pygam/pygam.py
@@ -2,322 +2,21 @@
 
 from __future__ import division
 from collections import defaultdict, OrderedDict
+from copy import deepcopy
 
 import numpy as np
-from numpy import random as rng
 import scipy as sp
-from sklearn.base import BaseEstimator
-from progressbar import ProgressBar
+from scipy import stats
 
-from copy import deepcopy
+from core import Core
+from penalties import cont_P, cat_P
+from distributions import NormalDist, BinomialDist
+from links import IdentityLink, LogitLink
+from callbacks import Deviance, Diffs, Accuracy, validate_callback
+from utils import check_dtype_, check_y, print_data, gen_knots, b_spline_basis
 
 
 EPS = np.finfo(np.float64).eps # machine epsilon
-
-def check_dtype_(X):
-    jitter = np.random.randn(X.shape[0])
-    dtypes_ = []
-    for feat in X.T:
-        dtype = feat.dtype.type
-        assert issubclass(dtype, (np.int, np.float)), 'data must be discrete or continuous valued'
-
-        if issubclass(dtype, np.int) or (len(np.unique(feat)) != len(np.unique(feat + jitter))):
-            assert (np.max(feat) - np.min(feat)) == (len(np.unique(feat)) - 1), 'k categories must be mapped to integers in [0, k-1] interval'
-            dtypes_.append(np.int)
-            continue
-
-        if issubclass(dtype, np.float):
-            dtypes_.append(np.float)
-            continue
-    return dtypes_
-
-def check_y(y, link, dist):
-    y = np.ravel(y)
-    assert np.all(~(np.isnan(link.link(y, dist)))), 'y data is not in domain of link function'
-    return y
-
-
-def nice_repr(name, param_kvs, line_width=30, line_offset=5, decimals=3):
-    """
-    tool to do a nice repr of a class.
-
-    Parameters
-    ----------
-    name : str
-        class name
-    param_kvs : dict
-        dict containing class parameters names as keys,
-        and the corresponding values as values
-    line_width : int
-        desired maximum line width.
-        default: 30
-    line_offset : int
-        desired offset for new lines
-        default: 5
-    decimals : int
-        number of decimal places to keep for float values
-        default: 3
-
-    Returns
-    -------
-    out : str
-        nicely formatted repr of class instance
-    """
-    if len(param_kvs) == 0:
-        # if the object has no params it's easy
-        return '%s()' % name
-
-    param_kvs = param_kvs[::-1]
-    out = ''
-    current_line = name + '('
-    while len(param_kvs) > 0:
-        k, v = param_kvs.pop()
-        if issubclass(v.__class__, (float, np.ndarray)):
-            # round the floats first
-            v = round_to_n_decimal_places(v, n=decimals)
-            param = '{}={},'.format(k, str(v))
-        else:
-            param = '{}={},'.format(k, repr(v))
-        if len(current_line + param) <= line_width:
-            current_line += param
-        else:
-            out += current_line + '\n'
-            current_line = ' '*line_offset + param
-
-        if len(current_line) < line_width and len(param_kvs) > 0:
-            current_line += ' '
-
-    out += current_line[:-1] # remove trailing comma
-    out += ')'
-    return out
-
-
-def round_to_n_decimal_places(array, n=3):
-    """
-    tool to keep round a float to n decimal places.
-
-    n=3 by default
-    """
-    # check if in scientific notation
-    if issubclass(array.__class__, float) and '%.e'%array == str(array):
-        return array # do nothing
-
-    shape = np.shape(array)
-    return ((np.atleast_1d(array) * 10**n).round().astype('int') / (10.**n)).reshape(shape)
-
-
-def print_data(data_dict, width=-5, keep_decimals=3, fill=' ', title=None):
-    """
-    tool to print a dictionary with a nice formatting
-
-    Parameters:
-    -----------
-    data_dict:
-        dict. Dictionary to be printed.
-    width:
-        int. Desired total line width.
-        A negative value will fill to minimum required width + neg(width)
-        default: -5
-    keep_decimals:
-        int. number of decimal places to keep:
-        default: 3
-    fill:
-        string. the character to fill between keys and values.
-        Must have length 1.
-        default: ' '
-    title:
-        string.
-        default: None
-    """
-
-    # find max length
-    keys = np.array(data_dict.keys(), dtype='str')
-    values = round_to_n_decimal_places(np.array(data_dict.values())).astype('str')
-    M = max([len(k + v) for k, v in zip(keys, values)])
-
-    if width < 0:
-        # this is for a dynamic filling.
-        # fill to minimum required width + neg(width)
-        width = M - width
-
-    assert M < width, 'desired width is {}, but max data length is {}'.format(width, M)
-
-    fill = str(fill)
-    assert len(fill) == 1, 'fill must contain exactly one symbol'
-
-    if title is not None:
-        print(title)
-        print('-' * width)
-    for k, v in zip(keys, values):
-        nk = len(k)
-        nv = len(v)
-        filler = fill*(width - nk - nv)
-        print(k + filler + v)
-
-
-def gen_knots(data, dtype, n_knots=10, add_boundaries=False):
-        """
-        generate knots from data quantiles
-
-        for discrete data, assumes k categories in [0, k-1] interval
-        """
-        assert dtype in [np.int, np.float], 'unsupported dtype'
-        if dtype == np.int:
-            knots = np.r_[np.min(data) - 0.5, np.unique(data) + 0.5]
-        else:
-            knots = np.percentile(data, np.linspace(0,100, n_knots+2))
-
-        if add_boundaries:
-            return knots
-        return knots[1:-1]
-
-def b_spline_basis(x, boundary_knots, order=4, sparse=True):
-    """
-    generate b-spline basis using De Boor recursion
-    """
-    x = np.atleast_2d(x).T
-    aug_knots = np.r_[boundary_knots.min() * np.ones(order-1), np.sort(boundary_knots), boundary_knots.max() * np.ones(order-1)]
-
-    bases = (x >= aug_knots[:-1]).astype(np.int) * (x < aug_knots[1:]).astype(np.int) # haar bases
-    bases[(x >= aug_knots[-1])[:,0], -order] = 1 # want the last basis function extend past the boundary
-    bases[(x < aug_knots[0])[:,0], order] = 1
-
-    maxi = len(aug_knots) - 1
-
-    # do recursion from Hastie et al.
-    for m in range(2, order + 1):
-        maxi -= 1
-        maskleft = aug_knots[m-1:maxi+m-1] == aug_knots[:maxi] # bookkeeping to avoid div by 0
-        maskright = aug_knots[m:maxi+m] == aug_knots[1:maxi+1]
-
-        left = ((x - aug_knots[:maxi]) / (aug_knots[m-1:maxi+m-1] - aug_knots[:maxi])) * bases[:,:maxi]
-        left[:,maskleft] = 0.
-
-        right = ((aug_knots[m:maxi+m]-x) / (aug_knots[m:maxi+m] - aug_knots[1:maxi+1])) * bases[:,1:maxi+1]
-        right[:,maskright] = 0.
-
-        bases = left + right
-
-    if sparse:
-        return sp.sparse.csc_matrix(bases)
-
-    return bases
-
-def ylogydu(y, u):
-    """tool to give desired output for the limit as y -> 0, which is 0"""
-    mask = (np.atleast_1d(y)!=0.)
-    out = np.zeros_like(u)
-    out[mask] = y[mask] * np.log(y[mask] / u[mask])
-    return out
-
-
-class Core(object):
-    """
-    core class
-
-    comes loaded with useful methods
-    """
-    def __init__(self, name=None, line_width=70, line_offset=3):
-        self._name = name
-        self._line_width = line_width
-        self._line_offset = line_offset
-        self._exclude = []
-
-    def __str__(self):
-        if self._name is None:
-            return self.__repr__()
-        return self._name
-
-    def __repr__(self):
-        name = self.__class__.__name__
-        param_kvs = [(k,v) for k,v in self.get_params().iteritems()]
-
-        return nice_repr(name, param_kvs, line_width=self._line_width, line_offset=self._line_offset)
-
-    def get_params(self):
-        return dict([(k,v) for k,v in self.__dict__.iteritems() if k[0]!='_' and (k not in self._exclude)])
-
-    def set_params(self, **parameters):
-        param_names = self.get_params().keys()
-        for parameter, value in parameters.items():
-            if parameter in param_names:
-                setattr(self, parameter, value)
-        return self
-
-class Distribution(Core):
-    """
-    base distribution class
-    """
-    def __init__(self, name=None, scale=None):
-        self.scale = scale
-        self._known_scale = self.scale is not None
-        super(Distribution, self).__init__(name=name)
-
-    def phi(self, y, mu, edof):
-        """
-        GLM scale parameter.
-        for Binomial and Poisson families this is unity
-        for Normal family this is variance
-        """
-        if self._known_scale:
-            return self.scale
-        else:
-            return np.sum(self.V(mu**-1) * (y - mu)**2) / (len(mu) - edof)
-
-class NormalDist(Distribution):
-    """
-    Normal Distribution
-    """
-    def __init__(self, scale=None):
-        super(NormalDist, self).__init__(name='normal', scale=scale)
-
-    def pdf(self, y, mu):
-        return np.exp(-(y - mu)**2/(2*self.scale)) / (self.scale * 2 * np.pi)**0.5
-
-    def V(self, mu):
-        """glm Variance function"""
-        return np.ones_like(mu)
-
-    def deviance(self, y, mu, scaled=True):
-        """
-        model deviance
-
-        for a gaussian linear model, this is equal to the SSE
-        """
-        dev = ((y - mu)**2).sum()
-        if scaled:
-            return dev / self.scale
-        return dev
-
-class BinomialDist(Distribution):
-    """
-    Binomial Distribution
-    """
-    def __init__(self, levels=1):
-        if levels is None:
-            levels = 1
-        self.levels = levels
-        super(BinomialDist, self).__init__(name='binomial', scale=1.)
-        self._exclude.append('scale')
-
-    def pdf(self, y, mu):
-        n = self.levels
-        return (sp.misc.comb(n, y) * (mu / n)**y * (1 - (mu / n))**(n - y))
-
-    def V(self, mu):
-        """glm Variance function"""
-        return mu * (1 - mu/self.levels)
-
-    def deviance(self, y, mu, scaled=True):
-        """
-        model deviance
-
-        for a bernoulli logistic model, this is equal to the twice the negative loglikelihod.
-        """
-        dev = 2 * (ylogydu(y, mu) + ylogydu(self.levels - y, self.levels-mu)).sum()
-        if scaled:
-            return dev / self.scale
-        return dev
 
 
 DISTRIBUTIONS = {'normal': NormalDist,
@@ -327,150 +26,12 @@ DISTRIBUTIONS = {'normal': NormalDist,
                  'inv_gaussian': None
                  }
 
-class Link(Core):
-    def __init__(self, name=None):
-        super(Link, self).__init__(name=name)
-
-class IdentityLink(Link):
-    def __init__(self):
-        super(IdentityLink, self).__init__(name='identity')
-
-    def link(self, mu, dist):
-        """
-        glm link function
-        this is useful for going from mu to the linear prediction
-        """
-        return mu
-
-    def mu(self, lp, dist):
-        """
-        glm mean ie inverse of link function
-        """
-        return lp
-
-    def gradient(self, mu, dist):
-        """
-        derivative of the linear prediction wrt mu
-        """
-        return np.ones_like(mu)
-
-class LogitLink(Link):
-    def __init__(self):
-        super(LogitLink, self).__init__(name='logit')
-
-    def link(self, mu, dist):
-        """
-        glm link function
-        this is useful for going from mu to the linear prediction
-        """
-        return np.log(mu / (dist.levels - mu))
-
-    def mu(self, lp, dist):
-        """
-        glm mean ie inverse of link function
-        for classification this is the prediction probabilities
-        """
-        elp = np.exp(lp)
-        return dist.levels * elp / (elp + 1)
-
-    def gradient(self, mu, dist):
-        """
-        derivative of the linear prediction wrt mu
-        """
-        return dist.levels/(mu*(dist.levels - mu))
-
-
 LINK_FUNCTIONS = {'identity': IdentityLink,
                   'log': None,
                   'logit': LogitLink,
                   'inverse': None,
                   'inv_squared': None
                   }
-
-# Penalty Matrix Generators
-def cont_P(n, diff_order=1):
-    """
-    builds a default proto-penalty matrix for P-Splines for continuous features.
-    penalizes the squared differences between adjacent basis coefficients.
-    """
-    if n==1:
-        return sp.sparse.csc_matrix(0.) # no second order derivative for constant functions
-    D = np.diff(np.eye(n), n=diff_order)
-    return sp.sparse.csc_matrix(D.dot(D.T))
-
-def cat_P(n):
-    """
-    builds a default proto-penalty matrix for P-Splines for categorical features.
-    penalizes the squared value of each basis coefficient.
-    """
-    return sp.sparse.csc_matrix(np.eye(n))
-
-# CallBacks
-def validate_callback_data(method):
-    def method_wrapper(*args, **kwargs):
-        expected = method.__code__.co_varnames
-
-        # rename curret gam object
-        if 'self' in kwargs:
-            gam = kwargs['self']
-            del(kwargs['self'])
-            kwargs['gam'] = gam
-
-        # loop once to check any missing
-        missing = []
-        for e in expected:
-            if e == 'self':
-                continue
-            if e not in kwargs:
-                missing.append(e)
-        assert len(missing) == 0, 'CallBack cannot reference: {}'.format(', '.join(missing))
-
-        # loop again to extract desired
-        kwargs_subset = {}
-        for e in expected:
-            if e == 'self':
-                continue
-            kwargs_subset[e] = kwargs[e]
-
-        return method(*args, **kwargs_subset)
-
-    return method_wrapper
-
-def validate_callback(callback):
-    if not(hasattr(callback, '_validated')) or callback._validated == False:
-        assert hasattr(callback, 'on_loop_start') or hasattr(callback, 'on_loop_end'), 'callback must have `on_loop_start` or `on_loop_end` method'
-        if hasattr(callback, 'on_loop_start'):
-            setattr(callback, 'on_loop_start', validate_callback_data(callback.on_loop_start))
-        if hasattr(callback, 'on_loop_end'):
-            setattr(callback, 'on_loop_end', validate_callback_data(callback.on_loop_end))
-        setattr(callback, '_validated', True)
-    return callback
-
-
-class CallBack(Core):
-    def __init__(self, name):
-        super(CallBack, self).__init__(name=name)
-
-@validate_callback
-class Deviance(CallBack):
-    def __init__(self):
-        super(Deviance, self).__init__(name='deviance')
-    def on_loop_start(self, gam, y, mu):
-        return gam.distribution.deviance(y=y, mu=mu, scaled=False)
-
-@validate_callback
-class Accuracy(CallBack):
-    def __init__(self):
-        super(Accuracy, self).__init__(name='accuracy')
-    def on_loop_start(self, y, mu):
-        return np.mean(y == (mu>0.5))
-
-@validate_callback
-class Diffs(CallBack):
-    def __init__(self):
-        super(Diffs, self).__init__(name='diffs')
-    def on_loop_end(self, diff):
-        return diff
 
 CALLBACKS = {'deviance': Deviance,
              'diffs': Diffs,
